feat(core): region_assignments O(N) + clinical dashboard + ParlaMint B8
Core:
- region_assignments(level, filter) — O(N) single-pass scan for all regions
- Temporal filtering support for region_assignments
- 4 new tests, fix all clippy warnings across codebase

Notebooks:
- B3 clinical dashboard with centered DSM-5 anchor projection
- B7 political polarization: fixed timestamps and counterfactual
- B8 ParlaMint-ES: real Spanish Parliament data (32K speeches, 2015-2023)

Docs:
- RFC-012: parallel HNSW construction + embedding space centering

+++ crates/cvx-core/src/types/episode_encoding.rs
@@ -29,7 +29,7 @@ pub fn encode_entity_id(episode_id: u64, step_index: u32) -> u64 {
         "episode_id {episode_id} exceeds 48-bit limit"
     );
     assert!(
-        step_index <= MAX_STEP_INDEX as u32,
+        step_index <= MAX_STEP_INDEX,
         "step_index {step_index} exceeds 16-bit limit"
     );
     (episode_id << 16) | (step_index as u64)
@@ -53,7 +53,7 @@ pub fn episode_start(episode_id: u64) -> u64 {
 /// `[start, end]` inclusive. Use with `TemporalFilter` or entity lookups.
 pub fn episode_range(episode_id: u64) -> (u64, u64) {
     let start = encode_entity_id(episode_id, 0);
-    let end = encode_entity_id(episode_id, MAX_STEP_INDEX as u32);
+    let end = encode_entity_id(episode_id, MAX_STEP_INDEX);
     (start, end)
 }
 
@@ -87,7 +87,7 @@ mod tests {
         assert_eq!(ep_s, 42);
         assert_eq!(st_s, 0);
         assert_eq!(ep_e, 42);
-        assert_eq!(st_e, MAX_STEP_INDEX as u32);
+        assert_eq!(st_e, MAX_STEP_INDEX);
     }
 
     #[test]
@@ -99,10 +99,10 @@ mod tests {
 
     #[test]
     fn max_values() {
-        let id = encode_entity_id(MAX_EPISODE_ID, MAX_STEP_INDEX as u32);
+        let id = encode_entity_id(MAX_EPISODE_ID, MAX_STEP_INDEX);
         let (ep, st) = decode_entity_id(id);
         assert_eq!(ep, MAX_EPISODE_ID);
-        assert_eq!(st, MAX_STEP_INDEX as u32);
+        assert_eq!(st, MAX_STEP_INDEX);
     }
 
     #[test]
@@ -114,6 +114,6 @@ mod tests {
     #[test]
     #[should_panic(expected = "exceeds 16-bit")]
     fn step_overflow_panics() {
-        encode_entity_id(0, MAX_STEP_INDEX as u32 + 1);
+        encode_entity_id(0, MAX_STEP_INDEX + 1);
     }
 }
